refactor: changed utils infrastructure
